Download webpage
Added the functionality to download a webpage. A consequence to
this code is that the program could crash if there is an error
downloading. I decided to not use a try/except clause because
in order to stop the program I would need to reraise the
exception or call system exit. Unsure what the best practice is
in this situation so I let it crash rather than redundantly force
it to crash.