Parse webpage for desired data
The data is stored into a list. Each element contains both the title
and the author. This will require regex for placing in proper database
cells. I chose to not scrape the data as separate elements of title
and author because one of the books does not have an author. This
causes the list to be offset (when intending to iterate over the list
and using data[i] as title and data[i+1] as author.)